requirements.txt
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
APScheduler==3.10.4
uvicorn==0.24.0.post1
fastapi==0.104.1